Add files via upload
The purpose of this analysis is to find which days of week does the LinkedIn pages of US compnies experience maximum and minimum number of increase in followers. This tells us the behaviors of LinkedIn users, that is what time of the day they are mostly active on linked in. This gives insight to companies on what days of the week people arre more likely to visit their LinkedIn page. Also this is a good metric for LinkedIn company as it better understand the variation of visitors' traffick with the days of the week.
We will use a data that contains the number of LinkedIn followers each day for thousands of US companies. The data was collected from 2015 - 2018. Data is not publicly available. If you need it you can contact me via Github.